TxHash component: rename ui/Hash to ui/ShortenedHash
