+++ js/src/ui/TxHash/txHash.js
@@ -21,7 +21,7 @@ import { bindActionCreators } from 'redux';
 import { LinearProgress } from 'material-ui';
 
 import { txLink } from '../../3rdparty/etherscan/links';
-import Hash from '../Hash';
+import ShortenedHash from '../ShortenedHash';
 
 import styles from './txHash.css';
 
@@ -67,7 +67,7 @@ class TxHash extends Component {
 
     const link = (
       <a href={ txLink(hash, isTest) } target='_blank'>
-        <Hash data={ hash } />
+        <ShortenedHash data={ hash } />
       </a>
     );
